cast

+++ server/src/group.cpp
@@ -16,7 +16,7 @@ void Group::removeUser(const std::string_view userName) {
 }
 
 int Group::postMessage(const std::string_view name, const std::string_view subject, const std::string_view content) {
-  int msgId = m_messages.size();
+  int msgId = (int)m_messages.size();
   Message newMessage = { msgId, name.data(), subject.data(), content.data() };
   return msgId;
 }
